+++ edu_project_vue/src/components/home/PostCardList.vue
@@ -569,6 +569,7 @@ const openImagePreview = (post: Post): void => {
   margin-bottom: 14px;
   display: -webkit-box;
   -webkit-line-clamp: 3;
+  line-clamp: 3;
   -webkit-box-orient: vertical;
   overflow: hidden;
 }