+++ edu_project_vue/src/views/admin/Reports.vue
@@ -540,6 +540,7 @@ onMounted(() => {
 .reason-text {
   display: -webkit-box;
   -webkit-line-clamp: 2;
+  line-clamp: 2;
   -webkit-box-orient: vertical;
   overflow: hidden;
 }