+++ edu_project_vue/src/views/user/UserProfile.vue
@@ -721,6 +721,7 @@ onMounted(() => {
   margin-bottom: 10px;
   display: -webkit-box;
   -webkit-line-clamp: 2;
+  line-clamp: 2;
   -webkit-box-orient: vertical;
   overflow: hidden;
 }